+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+New:
+
 - Introduce `logger` configuration's option to debug requests/responses.
 
   ```ruby
@@ -11,6 +13,10 @@ All notable changes to this project will be documented in this file.
     config.logger = Logger.new(STDOUT)
   end
   ```
+
+Fixes:
+
+- Add support for simple error structure.
 
 ## 0.0.1 (2016-02-03)